requirements.txt
streamlit
pandas==1.5.3
openai
google-cloud-aiplatform
anthropic
sqlalchemy
plotly
python-dotenv
requests
numpy==1.23.5
scipy
astor==0.8.1
duckdb==1.0.0
faker==19.12.0
matplotlib==3.7.1
sqlglot[rs]==25.0.3
torch==2.4.1
pillow==10.1.0
pandasai
mistralai